requirements.txt
flask==2.2.3
flask_session==0.4.0
spotipy==2.22.1
python-dotenv==1.0.0 